Rename image options in layout package for consistency with remote & local packages
