+++ layout/write.go
@@ -22,8 +22,8 @@ import (
 // The returned image will return layers when Layers(), LayerByDiffID(), or LayerByDigest() are called,
 // but the returned layer will error when DiffID(), Compressed(), or Uncompressed() are called.
 // This is useful when we need to satisfy the v1.Image interface but do not need to access any layers.
-func (l Path) AppendImage(img v1.Image, ops ...Option) error {
-	o := &options{}
+func (l Path) AppendImage(img v1.Image, ops ...AppendOption) error {
+	o := &appendOptions{}
 	for _, op := range ops {
 		op(o)
 	}
